requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pillow==9.4.0
graphene-django==3.0.0